+++ examples/resources/aiven_opensearch_security_plugin_config/import.sh
@@ -1,0 +1,1 @@
+terraform import aiven_opensearch_security_plugin_config.foo project/service_name